openrsync: avoid exposure via symlink race in the sender
This is actually more serious than the race depicted in the radar, because
the openrsync sender doesn't maintain a list of flist entries that are valid
to request and doesn't validate the flist stat when it enqueues the request.
Thus, the race isn't actually even necessary; a malicious client could just
request a symlink, and we'd happily follow it and send the contents.