openrsync: fix interaction with --copy-links
--copy-links in its current incarnation is implemented by just dereferencing
the symlink at open(2) time, so we can't apply O_NOFOLLOW globally.  We
should almost certainly not rely on the link dereferencing to the same file
twice, and we already sent metadata about the file over the wire.  Ideally,
we should dereference the link all of the way at flist generation time and
use that path for open(2).