openrsync: fix --list-only and the associated test
There is an obviously better implementation of --list-only and implied
--list-only than with a fake --dry-run: we're the receiver, we can simply
grab the flist and not bother requesting any files.  Doing so fixes
--list-only under protocol 29, since we would still send the flist index and
itemization flags on a dry run.

Fix the test to actually trigger this while we're here.  Previously, the
uploader would roll through all of the files and see that they're up to
date because the paths already exist in $PWD, then it would progress us
through the phases.  Create an empty dir to traverse into so that we can
ensure that we would have to request files if we implemented it incorrectly.

The test also now double-checks that we didn't actually create the dir that
was specified for transfer.