tests: openrsync: add a test to trigger our reallocation-segfault
In order to properly test this, we need to know how many flist entries we'll
allocate in a single run in order to trigger a reallocation when our
platform-specific bits add the AppleDouble file to the flist.  Export it
using a new --print-flist-chunk-size long opt since it's based on a certain
size worth of flist entries rather than # entries, and use that to populate
our src dir and reliably trigger reallocation.