openrsync: add more logging around directory handling
If we created the directory, make it clear that we did so with a log entry
to make it easier to trace through verbose output.  We also want to confirm
that we didn't transform the root directory somewhere along the way, since
we use the resulting dirfd in several *at() calls.