openrsync: kick back AT_RESOLVE_BENEATH violations in mkpath to the caller
Right now we just surface these as the mkdir() error, but that's not very
helpful because we couldn't confirm that the entry that existed is actually
a directory.  This means that we can return an EEXIST before we hit the leaf
that we actually wanted to create and think the directory exists, even if
we bailed out early.