openrsync: daemon: straighten out some log format confusion
An earlier implementation used opts->outformat because we didn't actually
have a clean separation of logging vs. stodut.  Now that we do, use the
correct format variable so that we get the transfer logging right.

While we're here, make sure all daemon option-sets have opts->daemon set to
be able to detect daemon usage anywhere we might need to, and the daemon
should set a logging session as it is an rsync server.