openrsync: move the client fd into the common role
We want to stop assuming that the log file has any relation to the client,
so push this detail into the role instead.  Clients will just set it to -1,
while the server will set it appropriately.  This is mostly important for
the daemon, which needs to log to a socket instead if it's bound for a
client.