openrsync: actually implement --quiet
This was previously implemented as, effectively, verbose=0, but there's more
to it than that if our logging/output stuff is all implemented correctly.

We don't have the session available in our logging path (and making it
available is quite ugly), so just make `quiet` global which has some
symmetry with `verbose`.  --quiet should prevent us from outputting these
non-error messages, but we still want to log them if logging was requested.