openrsync: don't crash on empty flists with --extended-attributes on macOS
Our platform_flist_received() implementation skips the last element since
it's only scanning for AppleDouble files and their matching file; the former
would have appeared before the latter in the flist, so if an AppleDouble
file appears last then something has perhaps gone wrong or we just can't do
anything about it.  We skipped the last element without confirming that we
actually had any to begin with, leading to an underflow/wrap and crash.