openrsync: avoid trying to use the wrong file type as a reference
open() of a symlink to a directory will succeed, but it's not really what
we want.  In this particular bug, we attempt to poll(2) a dirfd and fail for
obvious reasons.  Do an fstat(2) after we have it opened to confirm that it
is a regular file after all.

Note that there's a possibility here to have raced some other process
mucking with our destination, tree, too, but we rely on the EOF checksum
to fail and cause a file redo if such an event took place.