openrsync: do less metadata manipulation on existing files
In our previous versions, we do *all* of the metadata updates on files in
the destination before the transfer begins, then set metadata on the final
files after the transfer ends.  This means that we're often setting utimes,
ownership, etc. on files that will soon be getting replaced anyways, and
means we have more chances to fail where smb rsync would not.

Pare this down as much as we can; for files that are considered up-to-date,
this is where we have to do the final metadata update since we may or may
not be making a trip through the downloader with the file at all.  For other
files, we really only need to upgrade the file to readable so that we can
use it as a reference file for the transfer.  This fixes some spurious
errors we see in scenarios where the destination has mixed ownership, but
we're in the appropriate group to read the existing files.