openrsync: fix sending filenames with a prefix that matches the root
This is most easily reproducible if you add some dot-files to a directory,
cd into it, then attempt to `rsync . /some/dst`.  In our efforts to do some
sandboxing in the sender, we accidentally made it chop off the prefix of a
filename if it happened to also match the root directory's name.

Clearly that was not our intention, so let's fix it.  Ensure that if we chop
the root part of the path off, it was followed by a directory delimiter.