tests: fix-up the sparse tests to work more reliably
Stop checking non --sparse behavior: if holes are created, then it could
just be that the filesystem may be compressing the holes transparently for
us.  If the non --sparse rsync is doing hole checking, then that's just a
performance issue rather than a bug.

Switch to using a new mksparse tool for dealing with holes.  It has modes
to report holes on or to sparsify existing files, as well as to create new
sparse files of a shape defined via -d and -h arguments for data and hole
blocks respectively.  Instead of using external tooling to decide if we
created holes, we can just sparsify the reference file and compare the hole
reports from new and old.