openrsync: disable --sparse --inplace for now
We handle it incorrectly, and it didn't work until rsync 3.x anyways.  Thus,
stop explicitly testing for it for the time being and raise an error if one
tries the combination to avoid corrupting their file.  The correct move for
it to work is to make sure we actually use the appropriate hole-punching
mechanism for the platform.