openrsync: fix the first case in the test13c_sparse case
If the file ended in a hole, we would have lseek(2) past it but not actually
written anything after.  We need to actually extend the file to the correct
size.  We'll use the position we've been tracking in buf_copy() as a measure
to ensure that it's accurate; if we end up off in one direction or the
other, then comparisons in our test suite will fail.

Technically we only need to do this for --inplace and --sparse, but it does
not hurt to do it in the general case.