openrsync: sender: cleanup properly in all cases in rsync_sender()
The problem noted in this report is caused because we return immediately on
sender_get_iflags() or blk_recv() failures.  Our session stores a pointer
to the wbuf on the stack that we need to cleanup on the way out, so all
paths through rsync_sender() *must* jump to out so that we can properly
flush everything.