openrsync: fix copying complete file out of a partial directory
We'll resume a transfer out of a partial directory if the file is available,
which in some cases means that we'll just copy the file straight out because
it's complete.  In other cases, we'll continue building out the partial
file in place.

The problem with the former case is that it's wrong- the file is still in
the partial directory for a reason, so we should err on the side of caution
and run it through the normal update machinery rather than just accepting
the quick check.  This seems to match what smb rsync does.